pyproject.toml
[tool.poetry]
name = "postguardapi"
version = "0.1.0"
description = ""
authors = ["ImperatorNeron <katebarspk@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.4"}
scons = "^4.8.1"
poetry-dotenv = "^0.4.0"
alembic = "^1.13.3"
pydantic = "^2.9.2"
sqlalchemy = {extras = ["asyncpg"], version = "^2.0.36"}
asyncpg = "^0.30.0"
punq = "^0.7.0"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"
python-multipart = "^0.0.16"
openai = "^1.53.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
pre-commit = "3.8.0"
factory-boy = "^3.3.1"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[tool.pytest.ini_options]
pythonpath = [
    ".", "app",
]
asyncio_mode="auto"
asyncio_default_fixture_loop_scope = "session"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output=3
include_trailing_comma=true
line_length=120
lines_after_imports=2
lines_between_sections=1
lines_between_types=0
force_alphabetical_sort_within_sections=true
ensure_newline_before_comments=true
combine_as_imports=true
combine_star=true
skip_gitignore=true
balanced_wrapping=true
force_grid_wrap=2
src_paths=["app", ]
skip_glob=[
    ".git",
    ".idea",
    ".vscode",
    "venv",
    ".venv",
]

known_fastapi=["fastapi"]
known_first_party=["app"]
sections=[
    "FUTURE",
    "STDLIB",
    "FASTAPI",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]